OneHandMode Recents: fix NPE
In some situations, the system can set a fake empty app thumbnail because it's not able to take a correct screenshot. Thus if the user goes into one hand mode and open Recents, it will try to crop the thumbnail to scale it but the bitmap getWidth and getHeight methods will return a <0 value then throw an exception.
So we need to avoid the crop and just take the existing fake thumbnail.

Steps to replicate the issue:
- in normal mode, force kill the foreground app then open Recents and be sure the app thumbnail is an empty one;
- close Recents;
- go into One Hand Mode;
- open Recents and scroll them: SysUI FC.

Change-Id: Ic6507b29547fedf5426d62ec296d492ae5354b0a
Signed-off-by: Eliminater74 <eliminater74@gmail.com>